+++ src/main/java/com/fitnessapp/model/User.java
@@ -1,6 +1,5 @@
 package com.fitnessapp.model;
 
-
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -10,11 +9,10 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String fullName;
