+++ src/main/java/com/fitnessapp/repository/NutritionPlanRepository.java
@@ -3,6 +3,6 @@ package com.fitnessapp.repository;
 import com.fitnessapp.model.NutritionPlan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NutritionPlanRepository extends JpaRepository<NutritionPlan, Long> {
-    NutritionPlan findByUserId(Long userId);
+public interface NutritionPlanRepository extends JpaRepository<NutritionPlan, Integer> {
+    NutritionPlan findByUserId(Integer userId);
 }
