+++ src/main/java/com/fitnessapp/service/RecipeService.java
@@ -29,11 +29,11 @@ public class RecipeService {
         return recipeRepository.findByType(type);
     }
 
-    public Recipe getRecipeById(Long id) {
+    public Recipe getRecipeById(Integer id) {
         return recipeRepository.findById(id).orElse(null);
     }
 
-    public void deleteRecipe(Long id) {
+    public void deleteRecipe(Integer id) {
         recipeRepository.deleteById(id);
     }
 }
